+++ Q3E/src/main/jni/doom3/neo/framework/Common.cpp
@@ -41,6 +41,7 @@ extern bool Sys_InRenderThread(void);
 #define _HARM_DEBUG_MULTITHREAD
 #endif
 #endif
+#define HARM_ONLY_DETECT_SYS_MEMORY 1
 
 typedef enum {
 	ERP_NONE,
@@ -1637,6 +1638,7 @@ void Com_ExecMachineSpec_f(const idCmdArgs &args)
 		cvarSystem->SetCVarInteger("r_multiSamples", 0, CVAR_ARCHIVE);
 	}
 
+#if ! HARM_ONLY_DETECT_SYS_MEMORY
 	if (Sys_GetVideoRam() < 128) {
 		cvarSystem->SetCVarBool("image_ignoreHighQuality", true, CVAR_ARCHIVE);
 		cvarSystem->SetCVarInteger("image_downSize", 1, CVAR_ARCHIVE);
@@ -1646,6 +1648,7 @@ void Com_ExecMachineSpec_f(const idCmdArgs &args)
 		cvarSystem->SetCVarInteger("image_downSizeBump", 1, CVAR_ARCHIVE);
 		cvarSystem->SetCVarInteger("image_downSizeBumpLimit", 256, CVAR_ARCHIVE);
 	}
+#endif
 
 	if (Sys_GetSystemRam() < 512) {
 		cvarSystem->SetCVarBool("image_ignoreHighQuality", true, CVAR_ARCHIVE);
@@ -3166,6 +3169,21 @@ void idCommonLocal::SetMachineSpec(void)
 
 	Printf("Detected\n \t%.2f GHz CPU\n\t%i MB of System memory\n\t%i MB of Video memory\n\n", ghz, sysRam, vidRam);
 
+#if HARM_ONLY_DETECT_SYS_MEMORY
+    if (sysRam >= 1024) {
+        Printf("This system qualifies for Ultra quality!\n");
+        com_machineSpec.SetInteger(3);
+    } else if (sysRam >= 512) {
+        Printf("This system qualifies for High quality!\n");
+        com_machineSpec.SetInteger(2);
+    } else if (sysRam >= 384) {
+        Printf("This system qualifies for Medium quality.\n");
+        com_machineSpec.SetInteger(1);
+    } else {
+        Printf("This system qualifies for Low quality.\n");
+        com_machineSpec.SetInteger(0);
+    }
+#else
 	if (ghz >= 2.75f && vidRam >= 512 && sysRam >= 1024) {
 		Printf("This system qualifies for Ultra quality!\n");
 		com_machineSpec.SetInteger(3);
@@ -3181,6 +3199,7 @@ void idCommonLocal::SetMachineSpec(void)
 	}
 
 	com_videoRam.SetInteger(vidRam);
+#endif
 }
 
 /*